+++ jcpp/src/main/cpp/jcpp/io/JOutputStream.h
@@ -17,7 +17,7 @@ namespace jcpp{
         public:
             JOutputStream();
             static JClass* getClazz();
-            virtual void write(jbyte b)=0;
+            virtual void write(jint b)=0;
             virtual void write(jbyte b[], int off, int len);
             virtual void flush()=0;
             virtual void close()=0;
